MirRenderSurface/the-future-MirSurface: remove the MirBufferUsage parameter from the mir_render_surface_get_buffer_stream.
MirBufferUsage is being deprecated for 1.0, so we have to find someway to get rid of it from this fn before publishing. This proposes that get_buffer_stream() will only return a stream backed by software buffers. The mesa platform (in its egl driver) will use MirPresentationChain+MirBuffer. The android driver will use the android_egl extension provided by us for conversions within the hybris egl shim. The eglstream platform will provide its own streams, and not need MirBufferStreams.

The somewhat-contrived example of playground/egldiamond_render_surface.c was modified so that it actually needs a mir-platform EGL to run against. Previously, the example would make a 'hardware' buffer stream, and use that on any platform. Without a 'hardware' buffer stream anymore, the only reasonable thing to do is have the playground/ example that is used in mir-EGL platform development to require an actual mir-EGL platform to run against.