mirserver: Use the normal mir connection to send input to clients
Input Events are now sent through the connection and contain a window id to refer to the destination window. So all types of events occur on the client side in the order they were sent from the server. Additionally the conversion between android::InputMessage and MirEvent is avoided. This simplifies adding new event types or extending the content of existing input events. The client code needed only minor changes to not filter out MirInputEvent.