Disable predictive bypass, for a while, to work around LP: #1666372.
Predictive bypass works on the assumption that you are in phase with
the display vsync. But client-side vsync isn't really finished and is
still lacking phase correction. So the two optimizations are conflicting,
causing some stuttering.

Regrettably this will increase latency a bit (around 11ms) in some cases
for a while. But that's nothing compared to the 40-50ms improvement we
get from client-side vsync.

Eventually when phase correction is implemented in client-side vsync it
will be safe to re-enable predictive bypass.