Wait for the arrival of the input config message before attempting to propagate input device state events
The former informs the nested server about the available devices, the latter is supposed to synchronize the host server input state with the focused client (here the nested server). 

When launching the nested server there is a potential race between the initial input device state event and the input config. This problem is most visible when vt switching. Since the per-surface input fds have been removed, and all events are sent to the connection this race is harder to notice. Still the two messages may be sent by different threads in the server. This should be the last cause for the two prominent CI failures mentioned.