Ensure RealKMSOutputConfiguration stays in sync with actual hardware state.
For some reason we were going out of our way to ensure that the hardware configuration described
in RealKMSOutputConfiguration was out of sync with the actual hardware state.

This is not useful, and could result in shells rendering incorrectly, when the display configuration
tells them that the display buffers are a different size to what is being presented on screen.