Clear keystate and keymap on device removal
A client window may have a "default" keymap configured. So a keystate will be tracked for each keyboard device individually. Switching away from the vt of the mirserver will trigger a removal of all devices, switching back to the server will make those devices reappear.
      
If the keystate is not cleared on removal any pressed key will be stuck during a vt switch.