Remove Session::{create,destroy}_buffer.
There's no particularly good reason for the session to be involved in buffer management.
The fact that we don't construct buffers from IPC on each submission is entirely a
IPC optimisation, so manage that IPC optimisation in the frontend.

This, incidentally, makes it easier to support Wayland where the session is most
definitely *not* involved in buffer management.

It also makes supporting client-allocated buffers easier.