client: Don't set BufferStreamID to an invalid value when allocating a buffer.
We have code in mf::SessionMediator to check if a BufferStreamID is set on a allocation request.
This code never exectues, though, because for some reason MirConnection will always set
the BufferStreamID on an allocation request.

And if there isn't a reasonable BufferStreamID to set (ie: we're allocating a buffer via
mir_connection_allocate_buffer) then it sets -1, an invalid ID.