MultiMonitorArbiter: Drop the BufferMap.
The compositor is no longer in charge of sending buffers back to the client - this is now done
when the compositor releases its last reference.

So the MultiMonitorArbiter doesn't need access to the BufferMap to send buffers out.

This cascades through various tests which relied on sticking in a MockBufferMap to detect when
buffers are released. Instead, count references.