client: Don't wait for a BufferOperation::released event before freeing the client buffer.
The only reason to keep a track of buffers that the client has asked to be released but that
the server hasn't acknowledged the release of is so that we can handle ::update operations
on those buffers.

But there's nothing the client library can usefully do on that ::update operation, so instead
we can just ignore updates to buffers the client doesn't know about.

This *does* mean that we don't get debug messages when the client receives an update to a
buffer that it doesn't know about, but that's an acceptable loss.