buffer-release Executor: set the IPC thread free to roam!
Resolves a deadlock during shutdown the nested server.

During shutdown, a HostBuffer can outlive the SessionMediater, resulting in it
having sole ownership of the Executor. If a buffer response is processed during
this time, it will take the buffer's AtomicCallback::mutex lock and then invoke
the callback.

The callback results in the destruction of the Buffer, which causes ~AutoSendBuffer
to enque a buffer-release functor on the Executor.

Because the AutoSendBuffer is now the sole owner of the Executor,
~AutoSendBuffer *also* destroys the Exectutor.

The executor thread, meanwhile, has started processing the buffer-release
functor, invokes it, and then goes to destroy the functor.

The functor has lambda-captured a shared_ptr to the underlying HostBuffer,
and, since the SessionMediator has been destroyed, the functor now has
unique ownership of the HostBuffer.

So destroying the functor results in ~HostBuffer being called.
~HostBuffer calls mir_buffer_release, which, in turn, calls…
AtomicCallback::set_callback. Which tries to take
AtomicCallback::mutex, but that has been taken waaaaay back on
the original thread which resulted in ~AutoSendBuffer being called.

So, it's blocked waiting on AtomicCallback::mutex, and the thread
that *has* AtomicCallback::mutex is waiting on ~AutoSendBuffer,
which is waiting on ~ThreadedExecutor, which is waiting on
thread.join() for the thread that's waiting on AtomicCallback::mutex.

Urgh.

Instead, fully embrace the GC!

Share ownership of all the relevant objects between ThreadedExecutor
and the dispatch thread, and then .detach() the dispatch thread.
~ThreadedExecutor no longer needs to block waiting for the dispatch
thread to complete, and so deadlock avoided!