mf::BasicConnector: Don't hold naked boost::asio::io_service& in objects in std::shared_ptr<>.
The boost::asio::local::stream_protocol::socket keeps a hold of the io_service reference passed into its constructor;
it's thus incorrect for the boost::…::socket to outlive the io_service it's constructed with.

We weren't doing anything special to extend the lifetime of the io_service. It just so happened that all of the
components with shared ownership of the ::socket would be destroyed before BasicConnector, but there was no
mechanism that enforced that ordering.

So, once again, embrace the GC!

Use the aliasing shared_ptr contructor to take our socket and io_service and create a shared_ptr that owns
both the io_service and the socket.