ThreadExectuor: Oh my god, the bees!
Static storage duration of something containing std::mutex and std::condition_variable plays badly
with fork(), such as our esteemed testsuite uses.

It's also possible that Mir shells would fork() under certain circumstances, and it'd be nice if
we didn't make that break horribly.

So: register a pthread_atfork handler to quiesce the ThreadExecutor into a known-safe idle state,
and then restart it on the other side of the fork.