Wayland: Fix WlShmBuffer lifetime handling.
A single wl_buffer can be attached to multiple wl_surface-s or a single wl_surface
multiple times. We need to handle this case and only allocate a single WlShmSurface
to be shared between all the relevant BufferStreams, as we must only send a
wl_buffer.release() event after *all* users of the buffer have relinquished ownership.

Hang this on the existing DestructionShim infrastructure.