WlKeyboard: Reconstruct xkbcommon state in the frontend.
This is a hack which should be replaced by exposing the information we need
in the MirKeyEvent and in mf::Surface and…

This works for now by reconstructing xkbcommon state to serialise to the
Wayland client. Wayland clients will *not* have the correct keymap, however,
because we don't receive the MirKeymapEvent at an appropriate time.