DefaultInputDeviceHub: Use ->enqueue_with_guaranteed_execution() in remove_observer().
We need to run the remove operation on the ServerActionQueue, and we need to block
until the remove has completed.

But!

If the main loop has been stopped then the functor passed to ->enqueue() will
never be run, and we'll hang indefinitely.

Instead, use the new ->enqueue_with_guaranteed_execution(). This will run on
the main loop if it's running (and so won't race with any device adds), but
will be run immediately if the main loop is *not* running, resolving the deadlock.