miral-desktop: Drop --vt option.
logind tracks user sessions and manages the ACLs for various things based on there being
an active session on the current VT. One of the ACLs it manages is the /dev/dri/card* ACL.

When miral-shell starts and switches VT, logind notices that there isn't an active user
session associated with the current VT, and revokes all the user ACLs.

When using the Mir client API this is not a problem - miral-shell is running as root,
so the user ACLs aren't necessary, and the Mir protocol sends pre-authenticated
/dev/dri/card? file descriptors to the client, so clients don't need the user ACLs
either.

The mesa-private Wayland protocol used for mesa EGL support, however, sends just
the name - /dev/dri/card? - that the compositor is using and expects the client
to open that device node.

When miral-shell has switched VTs the client, run under a user account, doesn't
have permission to open the device node, so EGL fails.

It's surprisingly difficult to tell logind that there's a user session active on
a given VT, so just don't expose the option in the miral-desktop wrapper.

Fixes: https://bugs.launchpad.net/mir/+bug/1728574 (for some value of fixes)