WlShmBuffer: Copy Wayland buffer data on construction.
We need to do this eagerly because wl_shm_buffer_end_access(buffer) can attempt to send
a protocol error on buffer when SIGBUS is raised. This has to be done on the Wayland
thread.

Furthermore, wl_shm_buffer_begin_access()/wl_shm_buffer_end_access() need to be called
on the same thread, so we can't just punt wl_shm_buffer_end_access() to the eventloop.

Since we don't want the renderer thread to block waiting on the Wayland eventloop to process
the buffer we eagerly copy and then just upload the previously-copied data on bind().

As an optimisation this should be changed so that the glTexImage2D call is made on
construction, taking the (slow) texture upload off the renderer hotpath.