WlShellSurface: Handle the surface being destroyed before we send the configure message.
The bad buffer test causes a protocol error to be raised; this results in the client
being disconnected and all its resources being destroyed.

Since we punt the initial wl_shell_surface_send_configure request out to the eventloop
it might get run after the protocol error causes everything to be torn down, resulting
in a crash.

We probably need to factor this out into a generic event-sender, but for now this
lets us pass the wlcs tests.