Two wrongs don't make a right.
wayland_client: Calculate buffer size correctly.

We were setting the stride incorrectly in wayland_client, it is documented as the number of
bytes and we were using the number of pixels.

This resulted in the GL driver attempting to access outside the bounds of the data array,
and a segfault.

Fixes #78 (better)