mf::BufferStream: Don't pretend to be a scene::Surface.
mf::BufferStream::{add,remove}_observer(ms::SurfaceObserver) exists as an implementation detail
of BasicSurface, which needs to provide the SurfaceObserver::frame_posted observable.

Rather than pass the whole SurfaceObserver down to the BufferStream, just add a callback
to run on buffer submition to the BufferStream.

If need be this can be switched out to a full BufferStreamObserver, but it is not currently
needed.