Wayland: Use a null-BufferAllocator if we don't have a real one.
This is technically not *quite* following the Null Object Pattern; this BufferAllocator
will throw if anything tries to use it.

It is null in spirit, though - if we don't have a real BufferAllocator, or it fails
to bind to the Wayland display then there should be no way a client can submit a
non-shm buffer.