LogindConsoleServicesTest: Use g_dbus_connection_new_for_address().
g_bus_get() returns a *singleton* GDBusConnection to the system bus that's persistent
over the program lifetime.

This means that any issue in a test that damages the GDBusConnection state also destroys
all subsequent tests, which is not the ideal behaviour.

g_dbus_connection_new_for_address() returns a new GDBusConnection on each call, so
the tests are better isolated.