LogindConsoleServices: Fix the initial fd in acquire_device.
Because computing is built on hate and lies, the gdbus-codegen code silently fails
to handle calls with file-descriptors, returning instead the index into an (inaccessible)
array of file-descriptors. Since we only get one fd back, that index is always 0, which
just happens to be a perfectly valid fd. Just not the one we wanted.