mesa-kms: Detect when mesa is using the software renderer.
EGL-GBM recently gained the ability to use the llvm software renderer as
long as the DRM device supports dumb buffers.

The NVIDIA drm device *does* support dumb buffers, so mesa-kms would now
detect that it “fully” supports running on NVIDIA. However, this is
software-only support, and PlatformPriority::Best does not accurately
describe it ☺.

Instead, check the GL_RENDERER string in probe; the LLVMPipe software
renderer returns a string starting with “llvmpipe”, so match against
that and return PlatformPriority::Supported in that case.