eglstream-kms: Ensure a current EGLContext before constructing displays.
The Display construction requires making EGL calls which assume a current context.
In the past Mir has accidentally made an appropriate EGL context current on the
thread which constructs the displays before trying to construct the displays,
but that was never guaranteed.

Ensure that the appropriate context is current before trying to construct the
KMSDisplayConfiguration.