LogindConsoleServices: Handle having a Seat with no active session.
Mir requires that the login manager has already set up an active logind session for us,
so the correct thing to do when there isn't one is to refuse to construct a
LogindConsoleServices.

Unfortunately, the *way* logind says “there is no active session” is by returning an
object path of “/” for the ActiveSession property. This is a perfectly sensible
object path, so GDBus is happy to construct a proxy against it.

It only fails when trying to detect whether or not the Session is active, because
the “/” DBus object doesn't actually have an State property, surprisingly enough.
This results in  returning NULL, which doesn't strncmp well
against “active”…

Instead, check if the active session object path is reported to be “/”, and
throw a sensible exception if so.