ServerConfiguration: Introduce explicit --console-provider option.
For display managers like LightDM or GDM this can let them explicitly select how they
expect Mir to interact - via logind or to handle to console on its own.

For tests we can explicitly select the NullConsoleProvider, to speed up the tests and
to ensure the tests don't accidentally modify the system they're running on.