DefaultServerConfiguration: Maintain an owning ref on ConsoleServices.
We need to pass the_console_services() through to a number of components, some of which
(like platform probing) don't have anywhere to stash an owning reference. This means we
construct and destroy the_console_services() multiple times during server startup.

So, use a shared_ptr<> rather than a CachedPtr<> for console_services, leaving
DefaultServerConfiguration to have an owning ref and guaranteeing single-construction.

*: modulo if multiple threads try to access the_console_services() before it has been
constructed; however, CachedPtr is also not threadsafe in that case.