Disable "sanitize=vptr" as it leads to LOTS noise.
The toolchain generates RTTI for inline classes we use as interfaces.
At link time this information is included in each shared library that
uses the class. We do that a lot (one library uses the interface and
multiple libraries implement the interface).

The resulting multiple definitions are mostly harmless, any use of these
interfaces is detected as an error by ubsan.