LogindConsoleSession: Monitor the Active property, not the State property.
logind (frustratingly) doesn't actually send PropertyNotify when State changes, so
Mir was never being notified of being switched from/to.

Since we need to monitor the ‘semi-deprecated’ Active property anyway, just use
that rather than matching on State, since we don't know if State will be updated
at the same time, before, or after the Active PropertyNotify event and the Active
property is what we actually cared about in the first place!