example-server: Fix a locale initialization crash on musl
For std::locale() the C standard specifies that is valid to pass an empty
string as an argument for the locale name. This works as specified on musl.

On musl, passing an empty string for the locale name to
std::codecvt_byname() throws a runtime exception. It's uncertain whether
an empty string to std::codecvt_byname() should be valid or not for this
call.

Use the default locale "C" which works on musl too.