Pass a `Display const&` through to `create_buffer_allocator`.
In subsequent work the BufferAllocator needs to have an EGLContext
which shares objects with the compositor context. That context is
provided by the `Display`, so it's easiest if we just pass the `Display`
in to the constructor.

This obviously doesn't scale to heterogenous hybrid setups, where the
client allocator and the output compositor need not even share an
EGLDisplay (and, similarly, is specific to the GL renderer), but supporting
heterogeneous hybrid or non-GL rendering will require further interface
changes so don't try and preempt that here.

(Delete the eglstream-kms BufferAllocator test in this; it doesn't test
anything useful, and is now difficult to make pass)