WaylandExecutor: Don't leak on shutdown, harder.
Although the sanitizers didn't catch it there was a reference-cycle keeping objects alive:
If the wl_event_loop was destroyed before ~WaylandExecutor enqueued the cleanup work item
then a std::function containing a shared_ptr to State would be added to State::workqueue,
and so the state would never be freed.

Instead give the workqueue a terminated state which just drops any enqueue()d functor
on the floor, letting the destructor clean up the functor's resources.