mg::gl::Texture: Rework y_inverted() to make it self-describing.
“y_inverted()” might suggest to the reader that it needs to be flipped
in order to render the right way up; this is incorrect.

“y_inverted()” means that the texture is in the standard GL coördinate
space, with y decreasing with row, so you actually need to flip it only
if y_inverted() is *false*.

This confused me in the renderer, leading to everything being upside down,
so replace the bool with a descriptive enum.