Use Build-depends alternatives for evil.
Xenial doesn't have a new-enough NVIDIA driver to build eglstream-kms.
But we want to build Mir on Xenial, and want to do it out of this
git branch.
Do the simplest, most evil thing possible: add an alternative build
dependency on something harmless and rely on debian/rules to not
try building eglstream-kms on Xenial.

This will need to be stripped out before we try to push to the main
archive.