{eglstream,mesa}-kms: Ensure queried nodes actually support modesetting.
If the NVIDIA drivers are in use, but the experimental KMS support is not enabled,
the eglstream-kms driver would still bind to the device as the EGL implementation
claims to support everything we need.

Check explicitly for modesetting support on the node we're planning to use, so
we can fall back to something that might work. (Fixes: #672)