mesa-kms: Use EGL_EXT_platform_base.
This lets us explicitly ask for a GBM-backed EGL platform, rather than just hoping
the libEGL can correctly guess what we've handed in is a pointer to a `gbm_device`.

This is necessary for Mali support, as Mali's libEGL will *not* guess that what
you've passed in to `eglGetDisplay()` is a `gbm_device*`.