build: Add option to disable fatal compile warnings
`-Werror` is a fine default for development builds. It's not a good
choice for release builds, however, as it essentially makes the set of
warnings a compiler may emit part of the dependency chain.

Compilers *love* to add new warnings (and for good reason!), but this is
unhelpful when it prevents people from building a Mir release because
their compiler is too new.

Add the `MIR_FATAL_COMPILE_WARNINGS` CMake option, defaulted to ON.
Ideally this would default to ON in development builds, and OFF in
releases, but that can be followup work.

Distributions should set `-DMIR_FATAL_COMPILE_WARNINGS=OFF` in their
build scripts.