LTTNG: Void tracepoints don't need a dummy field.
We have a bunch of no-context tracepoints - things like compositor::started(), which
have no useful context to add to their bare emission.

LTTNG is quite happy to have tracepoints with no fields, but for some reason we were
creating tracepoints with a dummy integer field and always setting it to 0.

Just use a tracepoint with no fields instead.