Fail the build if libdrm's drmIsMaster is available but not used.
The drmIsMaster in libdrm was always *meant* to return bool, but
it turns out that there exists code that #include's <xf86drm.h>
but expects that “bool” is a valid identifier name.

We don't have that constraint, and it's mildly helpful to ensure
that we *actually* use the libdrm-provided drmIsMaster().