Wayland: Eagerly send buffer release events.
We want to use `wl_resource_post_event()` (which is what is generated for
`wl_buffer_send_release()` by wayland-scanner), as this marks the connection
as flushable (ie: we want the client to receive this event now).

We want the release events to make it to the client ASAP, so they can act
on it.