WaylandExecutor: Eagerly execute work when on Wayland thread.
This would appear to be a fairly minor optimisation, as the Wayland thread
was going to dispatch the Executor (and hence perform this work) before
going to sleep anyway.

However, because Wayland events are processed in-order, getting them
in the stream *now*, rather than after the rest of the events are processed/
sent, can make a reasonable difference in client behaviour.