clients: Make the mirclient socket optional.
We're deprecating libmirclient. A lot of platform interfaces are used only for
mirclient support, so it would be convenient for writing new platforms if
those interfaces could simply throw rather than returning a useless stub (and
having clients fail after connecting to the socket).

This adds a new, disabled by default, option `--enable-mirclient`.

When specified, Mir will listen on a mirclient socket as normal. If the server
uses a platform that throws rather than providing working mirclient interfaces
the server will fail to start up.

When not specified, the server will not provide a mirclient socket to listen
on, and platforms will not need to provide the mirclient interfaces.