wayland:OutputManager: Fix use-after-free with monitor hotplug
`wl_global_destroy()` only prevents the global from being published in future
and sends the `global_remove` signal; it does *not* destroy any `wl_resource*`
that result from clients binding to the global.

This meant that when a monitor unplug event caused us to destroy the `Output`
there would still be `wl_resource`s with pointers to the `resource_map` of
the now-defunct `Output`, just waiting for someone to call `wl_resource_destroy`
and cause `Output::resource_destructor` to merrily try and remove them from
`resource_map`.

This was somewhat hidden because we never actually *call* `wl_resource_destroy`
on a wl_output, because we (incorrectly) don't implement the client's destroy
request. Mostly it manifested as an ASan warning on shutdown, when libwayland
came to destroy all the client resources, but it could also have been
triggered by a client which had bound a `wl_output` quitting after a hot-unplug
event.

Closes: #671