Move graphics platform probing into the server.
I don't know why this was in mirplatform in the first place - it doesn't use anything
private to mirplatform, isn't something that a platform is likely to want to do,
and was only even used in libmirserver.

While we're here, split out “probe one module” from “select the best module from
this list” so we can run the probe on manually-selected graphics platforms, too.

This gives us a chance to show any diagnostics the platform would display during
probe (such as the eglstream-kms platform suggesting the appropriate kernel
parameters) and give a useful warning if the manually-selected platform doesn't
think it will work.

Closes: #1045