mgc::ShmBuffer: Switch to an in-memory ShmBuffer.
FileBackedShmBuffer is only backed by a file because the mirclient
API required that we be able to send fds over the wire.

We don't need to support the mirclient native-buffer-handles, so
just use a malloc'd allocation instead.

This allows us to use a single implementation for all our platforms. In
future we might add platform-specific SHM buffer implementations, but
that will only be needed to allow platform-specific optimisations.