mgc::ShmBuffer: Ensure an EGL context for destruction.
We can happily ensure that `bind()` and the like are only called with an active
GL context - it's a precondition of the call being useful at all.

It's much more difficult to ensure that the destructor is called with an active
GL context, particularly in the presence of pervasive `std::shared_ptr`.

Furthemore, if we share an EGL context across threads there's no guarantee that
it won't be bound *in another thread* when we come to `~ShmBuffer()`, putting
us back where we started.

So, have the platform give the `ShmBuffer` a way of running code with a
guaranteed-available GL context: the new `EGLContextDelegate`.

Fixes: #1125