bors: Don't delete merged branches.
Once you delete the base branch of a PR in GitHub it cannot
be retargetted. So, if you have a series of stacked PRs
and bors merges the base branch, deleting that merged branch
means all dependent PRs need to be resubmitted, losing any
commentary.

If bors does *not* delete the merged branch then the PRs can
have their base branch changed to master as the parent gets
merged.