Disable a couple of tests that try to frig with /dev/random.
There was once a problem with low entropy at system startup causing problems with unity-system-compositor, but we've not seen that for a long time, and this approach to testing doesn't work on Fedora rawhide.

I think these tests are now sufficiently low value to consider dropping them. But disabling ought to stop them hanging in CI.