mesa-kms: Handle EOPNOTSUPP from drmModeGetResources.
DRM will return `EOPNOTSUPP` when attempting to access the KMS APIs on a device
node which does not support KMS.

We can hit this if `drmCheckModesettingSupported` is inconclusve (because, for example,
we're on an ARM device where the GPU has no PCI address).

This is the case on the Raspberry Pi 4. `drmCheckModesettingSupported` is inconclusive,
because of course it is, but the V3D DRM node does not support modesetting.