mir-smoke-test-runner: Start with Mir's eglinfo
For the moment we don't use the return value to PASS/FAIL the test, as
eglinfo doesn't support selecting the platform to test.

There's an upstream branch which supports dumping the info for a specific
platform; should that land we can then gate on the Wayland and X11 EGL
platforms being supported.

Even before that, the output can be helpful for understanding the rest of
the mir-smoke-test-runner output.