cmake: Fix mir_discover_external_gtests.
We *were* stringifying command + arguments and passing that as the command
to execute. For example, "wlcs --gtest_filter=foo" was being treated
as a single argument, resulting in the test failing to find a binary called
`wlcs --gtest_filter=foo` rather than finding a `wlcs` binary and passing
`--gtest_filter=foo` as an argument.