tests/LogindConsoleServices: Silence some irrelevant warning messages.
A number of these tests don't *require* all of the logind DBus interface to be present
to pass, but the implementation will (correctly) spew warnings into the logs when it
cannot call those interface methods.

Add the extra interface points to dbusmock to keep the code from spewing irrelevant
warning messages.