GraphicsBufferAllocator: Drop `alloc_buffer` interfaces.
These were used in the mirclient days to allocate buffers for clients. In the
Wayland world this is handled by the platform internally - the `GraphicsBufferAllocator`
gets a `wl_display` during setup, publishes whatever Wayland objects it needs,
and so can handle any client requests without further Frontend interaciton.

`GraphicsBufferAllocator` retains the `alloc_software_buffer` interface, as it
remains useful for the rest of the server to have access to CPU-accessible
buffer allocation facilities.

This requires a large number of superficial test changes, as by default
all our (mirclient-based) tests were passing `mir_buffer_usage_hardware`,
which no longer works. For the vast majority of the tests, simply changing
that to `mir_buffer_usage_software` continues to work.

The tests which specifically tested the `alloc_buffer` interface, and those
which tested client-visible aspects of it (like the scanout flag tests) are
removed.