frontend/wayland: Hoist SHM buffer stride check.
We were relying on the `WaylandAllocator`'s `buffer_from_shm` to send a
protocol error when the stride of a SHM buffer couldn't possibly be correct.

Unfortunately, we only actually did this frontend::WlShmBuffer, which is now
only used when a platform does *not* support Wayland itself.

Hoist the check up to before calling `buffer_from_shm`, so this works for
all platforms rather than just the test platform!