wlcs: Add a fails-on-success test for each expected-failure.
This is slightly more complicated than it might initially appear, as we want
to be able to list tests as expected-failure *before* we run against a `wlcs`
that contains the test.

So it's a two-step dance: check if the test exists, and then *only if it does*
expect it to fail.