Platform probing: Add explicit `hosted` probe value.
Nested platforms can be fully-featured, but we should prefer direct-access
platforms.

Add a `PlatformPriority::hosted` value to let us distinguish between
“a degraded-mode direct access platform” (eg: mesa-kms with software rendering)
and a nested platform (eg: X11), and update the X11 platform to use it.

(The name chosen is `hosted` rather than `nested` as that conflicts with the
mir::graphics::nested namespace)