graphics: Remove RenderingPlatform::make_ipc_operations()
This was purely to support mirclient IPC. Since we no longer need to support
mirclient IPC for real clients, and we've stubbed enough of this to make the
acceptance tests which incidentally need mirclient work, remove it from the
driver interface.

This requires removing the tests which test for properties of the IPC package
(which we no longer care about), and tests which test for properties of the
specific BufferPackage that is received client-side (which we also don't care
about).