Unbreak offscreen renderer.
It turns out we check for EGL version strictly equal to 1.4,
but Mesa has returned 1.5 for a while (and NVIDIA has returned
1.5 for*ever*).

There's no reason we won't work with newer minor EGL versions,
so don't fail if we find them.

I'm not sure how valuable this is, given that it's been broken
for *some time* and no-one has noticed, but we might as well
do the easy fix.