tests/MockEGL: Handle libepoxy `dlopen`ing libEGL.
libepoxy dynamically loads libEGL, so our mocks in mock_egl.cpp aren't
used by calls through epoxy. (epoxy *also* `dlopen`s libGL/GLES, but
that's not a problem for now).

Add a `dlopen` interposer to catch attempts to load `libEGL.so.1` and
resolve them to the test binary instead so that the mock entrypoints
are used.