mtd::Mock{GL,EGL}: Implement mocks in epoxy-copmatible fashion
libepoxy `#define`s the various EGL/GLES and extension entrypoints to point to its
resolver functions.

This is very convenient when *using* those entrypoints! You just call the function,
and it works!

This is *less* convenient when trying to mock out those entrypoints.

So, rather than trying to not `#include` epoxy in the mock implementations, just
do an `#undef`/declare-as-C/define dance for each of the entrypoints we want to mock.