gbm-kms: Fix DRM modifier usage.
`DRM_FORMAT_MOD_INVALID` is the “use the non-modifier API” flag. As
such, it never makes sense to pass `DRM_FORMAT_MOD_INVALID` to either
EGL or the KMS APIs.

This means we don't need an `optional<uint64_t>` to treat an unset
modifier as different to `DRM_FORMAT_MOD_INVALID`.