gbm-kms/linux-dmabuf: Support drivers without modifiers.
When the driver advertises the EGL modifier extension but does not
advertise any modifiers, this means that the driver has internal magic
to do the right thing with `DRM_FORMAT_MOD_INVALID`.

Thus, if the EGL does not advertise any modifiers for a format it
claims to support we should claim that it supports
`DRM_FORMAT_MOD_INVALID` for that format.