gbm-kms: Disable bypass by default.
This triggers a GPU OOM (issue #1801) on the Pi3 when used with default settings.
There may be due to us hanging on to buffers longer than they need to be, but
it's not clear yet.

Users who want the performance boost and are running on known hardware can
set `--bypass=true`, but by default use the safe option of `false`.