CMake: Modernise mirplatform target.
Modern CMake allows us to add dependencies to OBJECT libraries. They then
get their library dependencies *and* their header dependencies added to
whatever you link them to.

With a little helper to combine a bunch of OBJECT libraries into a
SHARED library, this allows us to make all of the OBJECT libraries that
make up `mirplatform` independent and automatically usable by simply
`target_link_libraries`ing.

Ideally we'd also be able to aggregate this, so we had a `mirplatform-obj`
for the unit-tests that got built into a `mirplatform` shared library.
Unfortunately I've not found a way to make that work properly, so
the `MIR_PLATFORM_OBJECTS` variable remains to list the OBJECT libraries
a unit-test testing `mirplatform` will need to link to.

This removes the necessity for the `MIR_PLATFORM_REFERENCES` variable,
and for any `include_directories()` calls required by linking to
platform objects.