Drop mirclient shared library completely.
Now that we're not using protobuf we no longer have a process-wide
singleton that'll `abort()` if we accidentally include it twice!

This means that we can just build mirclient as a static library,
have the relevant bits linked in directly to the consumers.

The last (user-visible) vestiges of the republic have been
swept away!