wayland::BufferCursorImage: Avoid locking a buffer indefinitely.
The `graphics::CursorImage` interface is much like an `mrs::Mapping`, in that
it allows arbitrary access to the data for as long as the `CursorImage` exists.

Unfortunately, this makes an `mrs::Mapping` a poor backing-store for a `CursorImage`,
as `Mapping`s necessarily lock the underlying buffer and are expected to be short
lived.

The most visible problem this caused was a deadlock; when a client set the cursor
to the *same* buffer as was currently used as a `BufferCursorImage`, it would
attempt to create a new `Mapping` from that buffer before destroying the old
`BufferCursorImage` which would ned to happen first in order to release the lock
that the new `Mapping` was trying to take…

Solve this by the time-honoured method of copying the data out of the buffer
on construction of the `BufferImage`.

It might separately be a good idea to allow multiple `Mapping`s of the same
buffer to exist simultaneously, but even if that worked it's not a good idea
to lock the buffer for so long.

Fixes: #2107